+++ src/config/mod.rs
@@ -1,6 +1,8 @@
 mod config;
+mod frequency;
 pub(crate) mod generators;
 mod user_config;
 
 pub use config::Config;
-pub use user_config::{Frequency, Generators};
+pub use frequency::Frequency;
+pub use generators::Generators;
